feat: use tree doctrine behavior for hierarchical entities
