+++ tests/App/Entity/Category.php
@@ -4,21 +4,21 @@ declare(strict_types=1);
 
 namespace whatwedo\CrudBundle\Tests\App\Entity;
 
-use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Validator\Constraints as Assert;
-use whatwedo\CrudBundle\Entity\HierarchicalEntityInterface;
-use whatwedo\CrudBundle\Entity\HierarchicalEntityTrait;
+use Gedmo\Mapping\Annotation as Gedmo;
+use Gedmo\Tree\Entity\Repository\NestedTreeRepository;
+use Gedmo\Tree\Traits\NestedSetEntity;
 
 /**
- * @ORM\Table(name="category")
- * @ORM\Entity(repositoryClass="whatwedo\CrudBundle\Tests\App\Repository\CategoryRepository")
- * @ORM\HasLifecycleCallbacks()
+ * @Gedmo\Tree(type="nested")
+ * @ORM\Entity(repositoryClass="Gedmo\Tree\Entity\Repository\NestedTreeRepository")
  */
-class Category implements HierarchicalEntityInterface
+#[Gedmo\Tree(type: 'nested')]
+#[ORM\Entity(repositoryClass: NestedTreeRepository::class)]
+class Category
 {
-    use HierarchicalEntityTrait;
     /**
      * @ORM\Column(type="integer")
      * @ORM\Id
@@ -28,25 +28,95 @@ class Category implements HierarchicalEntityInterface
 
     /**
      * @ORM\Column(type="string")
-     * @Assert\NotBlank
-     * @Assert\NotNull()
      */
     private ?string $name = null;
 
     /**
-     * @ORM\ManyToOne(targetEntity=Category::class, inversedBy="children")
+     * @var int|null
+     *
+     * @Gedmo\TreeLeft
+     * @ORM\Column(name="lft", type="integer")
      */
-    protected $parent = null;
+    #[Gedmo\TreeLeft]
+    #[ORM\Column(name: 'lft', type: Types::INTEGER)]
+    private $lft;
 
     /**
-     * @ORM\OneToMany(targetEntity=Category::class, mappedBy="parent")
+     * @var int|null
+     *
+     * @Gedmo\TreeLevel
+     * @ORM\Column(name="lvl", type="integer")
      */
-    protected  Collection $children;
+    #[Gedmo\TreeLevel]
+    #[ORM\Column(name: 'lvl', type: Types::INTEGER)]
+    private $lvl;
 
-    public function __construct()
+    /**
+     * @var int|null
+     *
+     * @Gedmo\TreeRight
+     * @ORM\Column(name="rgt", type="integer")
+     */
+    #[Gedmo\TreeRight]
+    #[ORM\Column(name: 'rgt', type: Types::INTEGER)]
+    private $rgt;
+
+    /**
+     * @var self|null
+     *
+     * @Gedmo\TreeRoot
+     * @ORM\ManyToOne(targetEntity="Category")
+     * @ORM\JoinColumn(name="tree_root", referencedColumnName="id", onDelete="CASCADE")
+     */
+    #[Gedmo\TreeRoot]
+    #[ORM\ManyToOne(targetEntity: Category::class)]
+    #[ORM\JoinColumn(name: 'tree_root', referencedColumnName: 'id', onDelete: 'CASCADE')]
+    private $root;
+
+    /**
+     * @var self|null
+     *
+     * @Gedmo\TreeParent
+     * @ORM\ManyToOne(targetEntity="Category", inversedBy="children")
+     * @ORM\JoinColumn(name="parent_id", referencedColumnName="id", onDelete="CASCADE")
+     */
+    #[Gedmo\TreeParent]
+    #[ORM\ManyToOne(targetEntity: Category::class, inversedBy: 'children')]
+    #[ORM\JoinColumn(name: 'parent_id', referencedColumnName: 'id', onDelete: 'CASCADE')]
+    private $parent;
+
+    /**
+     * @var Collection<int, Category>
+     *
+     * @ORM\OneToMany(targetEntity="Category", mappedBy="parent")
+     * @ORM\OrderBy({"lft" = "ASC"})
+     */
+    #[ORM\OneToMany(targetEntity: Category::class, mappedBy: 'parent')]
+    #[ORM\OrderBy(['lft' => 'ASC'])]
+    private $children;
+
+
+    public function getRoot(): ?self
     {
-        $this->children = new ArrayCollection();
+        return $this->root;
     }
+
+
+    public function getLevel(): int
+    {
+        return $this->lvl;
+    }
+
+    public function setParent(self $parent = null): void
+    {
+        $this->parent = $parent;
+    }
+
+    public function getParent(): ?self
+    {
+        return $this->parent;
+    }
+
 
     public function getId(): ?int
     {
@@ -65,6 +135,10 @@ class Category implements HierarchicalEntityInterface
         return $this;
     }
 
+    public function getChildren(): Collection
+    {
+        return $this->children;
+    }
 
     public function __toString(): string
     {
