+++ tests/App/config/packages/doctrine.yaml
@@ -1,29 +1,35 @@
 doctrine:
-  dbal:
-    default_connection: default
-    connections:
-      default:
-        # configure these for your database server
-        driver: 'pdo_sqlite'
-        charset: utf8mb4
-        default_table_options:
-          charset: utf8mb4
-          collate: utf8mb4_unicode_ci
-        url: '%env(resolve:DATABASE_URL)%'
-  orm:
-    default_entity_manager: default
-    auto_generate_proxy_classes: true
-    entity_managers:
-      default:
-        connection: default
-        naming_strategy: doctrine.orm.naming_strategy.underscore_number_aware
-        auto_mapping: true
-        mappings:
-          App:
-            is_bundle: false
-            type: annotation
-            dir: '%kernel.project_dir%/Entity'
-            prefix: 'whatwedo\CrudBundle\Tests\App\Entity'
-            alias: App
-    resolve_target_entities:
-        whatwedo\TableBundle\Entity\UserInterface: whatwedo\CrudBundle\Tests\App\Entity\Person
+    dbal:
+        default_connection: default
+        connections:
+            default:
+                # configure these for your database server
+                driver: 'pdo_sqlite'
+                charset: utf8mb4
+                default_table_options:
+                    charset: utf8mb4
+                    collate: utf8mb4_unicode_ci
+                url: '%env(resolve:DATABASE_URL)%'
+    orm:
+        default_entity_manager: default
+        auto_generate_proxy_classes: true
+        entity_managers:
+            default:
+                connection: default
+                naming_strategy: doctrine.orm.naming_strategy.underscore_number_aware
+                auto_mapping: true
+                mappings:
+                    App:
+                        is_bundle: false
+                        type: annotation
+                        dir: '%kernel.project_dir%/Entity'
+                        prefix: 'whatwedo\CrudBundle\Tests\App\Entity'
+                        alias: App
+
+                    tree:
+                        type: annotation # or attribute
+                        alias: Gedmo
+                        prefix: Gedmo\Tree\Entity
+                        dir: "%kernel.project_dir%/../../vendor/gedmo/doctrine-extensions/src/Tree/Entity"
+        resolve_target_entities:
+            whatwedo\TableBundle\Entity\UserInterface: whatwedo\CrudBundle\Tests\App\Entity\Person
