+++ tests/App/config/services.yaml
@@ -6,3 +6,12 @@ services:
         autowire: true
         autoconfigure: true
 
+    whatwedo\CrudBundle\Tests\App\Definition\:
+        resource: '../Definition'
+
+    gedmo.listener.tree:
+        class: Gedmo\Tree\TreeListener
+        tags:
+            - { name: doctrine.event_subscriber, connection: default }
+        calls:
+            - [ setAnnotationReader, [ "@annotation_reader" ] ]
